backend/requirements.txt
flask
Flask-CORS
google-cloud-firestore
gunicorn